requirements.txt
flask==3.1.3
mysql-connector-python==9.6.0
gunicorn==23.0.0
python-dotenv==1.0.0